requirements.txt
streamlit>=1.28.0
openai>=1.3.0
transformers>=4.35.0
torch>=2.1.0
sentence-transformers>=2.2.0
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
Pillow>=10.0.0

